graph build

+++ dflow_calc.cpp
@@ -30,10 +30,10 @@ public:
         {
             // iterate over all the neighbors of this particular node
             int node = p.first;
-            list <pair <int, int>> neighbour = p.second;
+            list <pair <int, double>> neighbour = p.second;
  
             cout << "Neighbors of: " << node << " are:\n";
-            pair<int, int> deps;
+            pair<int, double> deps;
             
             for(auto nbr : neighbour)
             {
